[FEAT] Separate format and other strings #2
To make it easier to skip other strings which might use `{}` (for example in
regexes) it separates those strings where `format` is called and every other
(apart from the docstrings which are already split).

It now uses the code 103 which was previously used for cases when both
implicit and explicit were used. Moved this into the 200 range using 205 as
this makes it possible to ignore all 2.6 related issues by just adding `P1`.

Also fixes the tests accordingly.