Allow Python 2.6 tests to fail
Since version 0.2.0 the script is using Python's formatter to analyse the
string which means that implicit parameters aren't detected in Python 2.6.

It should be still possible to use it on Python 2.6, just that any string with
implicit parameters will be handled incorrectly. But any error above P200 will
work on Python 2.6 compatible format strings.