[FEAT] Handle percent string format
Additionally to `str.format` it's possible to format strings using the
%-operator. This does actually check the format string if it is written
correctly as the string needs to be manually parsed.

It can only do some checks regarding the parameters as the type of the right
side is important and may not be known. At the moment it can only check the
tuples and dictionaries defined via the brackets (not via the functions).