Disallow failure of Python 2.6 builds again
The Python 2.6 build was allowed to fail, because it couldn't detect format
strings which use implicit parameters. It cannot just ignore P101-P103 results
as it'll also detect unused parameters (as P301) because it didn't detect that
it would be used on Python 2.7 and up.

So it just skips the tests which would usually fail.