Improve error message comparison
The error message comparison tests were using basic assertions and both
iterating over iterators. So it was only showing the first error and didn't
really compare both entries. It could also only report the first difference.

It now compares all reports with the expected reports and prints the
difference. Additionally it checks through the reports that are missing and
tries to find exactly one unexpected report which has only one different
attribute (like the line number) and will actually tell that it expected the
other values. That report won't show up as unexpected but instead together
with the missing one in that list.