Support including raw strings
Raw strings usually are regex expressions, which cannot really be checked
without also parsing the regex itself. So it by default will now ignore raw
strings but it can be enabled.

This also uses the new feature from Flake8 3.x to directly get the tokens of a
file without needing to open it. To support Flake8 2.x a wrapper is added
which manually creates the tokens.

The generic tests now also allow parameters for the different errors. For
example some errors only apply when errors with raw strings are enabled, so it
checks whether the errors are issued only when the setting is enabled.