Allow starargs not only at the end
With Python 3.5 the number and location of starargs (as well as the kwargs)
was relaxed. With 2b0b21c this was implemented but "required" the starargs to
be at the end. In the end most checks are disabled when starargs are used, so
the location shouldn't really matter, relaxing the condition.
Additionally it supports now any number of starargs.